formulas from DID moved
