+++ uds/translator/data_record_definitions/__init__.py
@@ -1,31 +1,33 @@
 """Data Records definitions for basic Translator."""
 
+from .conditional import (
+    CONDITIONAL_EVENT_TYPE_RECORD_09_2020,
+    DTC_SNAPSHOT_RECORDS_LIST_2013,
+    DTC_SNAPSHOT_RECORDS_LIST_2020,
+    DTC_STORED_DATA_RECORDS_LIST_2013,
+    DTC_STORED_DATA_RECORDS_LIST_2020,
+    DTCS_WITH_STATUSES_AND_EXTENDED_DATA_RECORDS_DATA_LIST,
+    EVENT_TYPE_RECORD_01,
+    EVENT_TYPE_RECORD_03_2013,
+    EVENT_TYPE_RECORD_03_2020,
+    EVENT_TYPE_RECORD_07_2013,
+    EVENT_TYPE_RECORD_07_2020,
+    EVENT_TYPE_RECORD_09_2020,
+)
 from .did import (
     DATA_FROM_DID_2013,
     DATA_FROM_DID_2020,
     DID_2013,
     DID_2020,
     DYNAMICALLY_DEFINED_DID,
-    EVENT_TYPE_RECORD_03_2013,
-    EVENT_TYPE_RECORD_03_2020,
-    EVENT_TYPE_RECORD_07_2013,
-    EVENT_TYPE_RECORD_07_2020,
     MULTIPLE_DID_2013,
     MULTIPLE_DID_2020,
     MULTIPLE_PERIODIC_DID,
     OPTIONAL_DYNAMICALLY_DEFINED_DID,
     OPTIONAL_MULTIPLE_PERIODIC_DID,
     OPTIONAL_PERIODIC_DID,
-    get_did_data_2013,
-    get_did_data_2020,
-    get_did_data_mask_2013,
-    get_did_data_mask_2020,
-    get_dids_2013,
-    get_dids_2020,
 )
-from .conditional import CONDITIONAL_EVENT_TYPE_RECORD_09_2020,    EVENT_TYPE_RECORD_01, EVENT_TYPE_RECORD_09_2020
 from .dtc import (
-
     DTC,
     DTC_AND_STATUS,
     DTC_COUNT,
@@ -39,16 +41,10 @@ from .dtc import (
     DTC_SEVERITY_AVAILABILITY_MASK,
     DTC_SEVERITY_MASK,
     DTC_SNAPSHOT_RECORD_NUMBER,
-    DTC_SNAPSHOT_RECORDS_LIST_2013,
-    DTC_SNAPSHOT_RECORDS_LIST_2020,
     DTC_STATUS,
     DTC_STATUS_AVAILABILITY_MASK,
     DTC_STATUS_MASK,
     DTC_STORED_DATA_RECORD_NUMBER,
-    DTC_STORED_DATA_RECORDS_LIST_2013,
-    DTC_STORED_DATA_RECORDS_LIST_2020,
-    DTCS_WITH_STATUSES_AND_EXTENDED_DATA_RECORDS_DATA_LIST,
-
     FAULT_DETECTION_COUNTER,
     GROUP_OF_DTC,
     MULTIPLE_DTC_AND_FAULT_DETECTION_COUNTERS,
