+++ uds/translator/data_record_definitions/did.py
@@ -2,18 +2,14 @@
 
 __all__ = [
     "DID_2020", "DID_2013",
+    "DID_DATA_MAPPING_2020", "DID_DATA_MAPPING_2013",
     "DYNAMICALLY_DEFINED_DID", "OPTIONAL_DYNAMICALLY_DEFINED_DID", "OPTIONAL_PERIODIC_DID",
     "MULTIPLE_DID_2020", "MULTIPLE_DID_2013",
     "MULTIPLE_PERIODIC_DID", "OPTIONAL_MULTIPLE_PERIODIC_DID",
     "DATA_FROM_DID_2013", "DATA_FROM_DID_2020",
-    "EVENT_TYPE_RECORD_03_2020", "EVENT_TYPE_RECORD_03_2013",
-    "EVENT_TYPE_RECORD_07_2020", "EVENT_TYPE_RECORD_07_2013",
-    "get_did_2020", "get_did_2013", "get_dids_2020", "get_dids_2013",
-    "get_did_data_2020", "get_did_data_2013", "get_did_data_mask_2020", "get_did_data_mask_2013",
-    "get_did_records_formula_2020", "get_did_records_formula_2013",
 ]
 
-from typing import Callable, List, Optional, Tuple, Union
+from typing import Dict
 
 from uds.utilities import (
     DID_BIT_LENGTH,
@@ -23,365 +19,17 @@ from uds.utilities import (
     PERIODIC_DID_OFFSET,
 )
 
-from ..data_record import (
-    AbstractDataRecord,
-    AliasMessageStructure,
-    ConditionalFormulaDataRecord,
-    LinearFormulaDataRecord,
-    MappingDataRecord,
-    RawDataRecord,
-)
-from .other import (
-    ACTIVE_DIAGNOSTIC_SESSION,
-    COMPARE_VALUE,
-    COMPARISON_LOGIC,
-    HYSTERESIS_VALUE,
-    LOCALIZATION,
-    RESERVED_BIT,
-)
+from ..data_record import AliasMessageStructure, LinearFormulaDataRecord, MappingDataRecord, RawDataRecord
+from .other import ACTIVE_DIAGNOSTIC_SESSION, RESERVED_BIT
 
-DID_DATA_MAPPING_2013 = {
+DID_DATA_MAPPING_2020: Dict[int, AliasMessageStructure] = {
     0xF186: (RESERVED_BIT, ACTIVE_DIAGNOSTIC_SESSION),
 }
-DID_DATA_MAPPING_2020 = {
-    0xF186: DID_DATA_MAPPING_2013[0xF186],
+"""DID values mapping (compatible with ISO 14229-1:2020) to DID data message structure."""
+DID_DATA_MAPPING_2013: Dict[int, AliasMessageStructure] = {
+    0xF186: DID_DATA_MAPPING_2020[0xF186],
 }
-
-
-def get_did_2013(name: str = "DID", optional: bool = False) -> MappingDataRecord:
-    """
-    Get DID Data Record compatible with ISO 14229-1:2013 version.
-
-    :param name: Name to assign to the Data Record.
-    :param optional: False if the Data Record presence is mandatory, True otherwise.
-
-    :return: Created DID Data Record.
-    """
-    return MappingDataRecord(name=name,
-                             length=DID_BIT_LENGTH,
-                             values_mapping=DID_MAPPING_2013,
-                             min_occurrences=0 if optional else 1,
-                             max_occurrences=1)
-
-
-def get_did_2020(name: str = "DID", optional: bool = False) -> MappingDataRecord:
-    """
-    Get DID Data Record compatible with ISO 14229-1:2020 version.
-
-    :param name: Name to assign to the Data Record.
-    :param optional: False if the Data Record presence is mandatory, True otherwise.
-
-    :return: Created DID Data Record.
-    """
-    return MappingDataRecord(name=name,
-                             length=DID_BIT_LENGTH,
-                             values_mapping=DID_MAPPING_2020,
-                             min_occurrences=0 if optional else 1,
-                             max_occurrences=1)
-
-
-def get_did_data_2013(name: str = "DID data") -> ConditionalFormulaDataRecord:
-    """
-    Get Conditional Data Record for DID data that is compatible with ISO 14229-1:2013 version.
-
-    :param name: Name for the Data Record that contains whole DID data.
-
-    :return: Conditional Data Record for DID data.
-    """
-    default_did_data = RawDataRecord(name=name,
-                                     length=8,
-                                     min_occurrences=1,
-                                     max_occurrences=None)
-
-    def _get_did_data(did: int) -> Tuple[RawDataRecord]:
-        data_records = DID_DATA_MAPPING_2013.get(did, None)
-        if data_records is None:
-            raise ValueError(f"No data structure defined for DID 0x{did:04X}.")
-        total_length = 0
-        for dr in data_records:
-            if not isinstance(dr, AbstractDataRecord) or not dr.fixed_total_length:
-                raise ValueError(f"Incorrectly defined data structure for DID 0x{did:04X}. "
-                                 f"Only fixed length data records are supported right now.")
-            total_length += dr.min_occurrences * dr.length
-        return (RawDataRecord(name=name,
-                              children=data_records,
-                              length=total_length,
-                              min_occurrences=1,
-                              max_occurrences=1),)
-
-    return ConditionalFormulaDataRecord(formula=_get_did_data,
-                                        default_message_continuation=[default_did_data])
-
-
-def get_did_data_2020(name: str = "DID data") -> ConditionalFormulaDataRecord:
-    """
-    Get Conditional Data Record for DID data that is compatible with ISO 14229-1:2020 version.
-
-    :param name: Name for the Data Record that contains whole DID data.
-
-    :return: Conditional Data Record for DID data.
-    """
-    default_did_data = RawDataRecord(name=name,
-                                     length=8,
-                                     min_occurrences=1,
-                                     max_occurrences=None)
-
-    def _get_did_data(did: int) -> Tuple[RawDataRecord]:
-        data_records = DID_DATA_MAPPING_2020.get(did, None)
-        if data_records is None:
-            raise ValueError(f"No data structure defined for DID 0x{did:04X}.")
-        total_length = 0
-        for dr in data_records:
-            if not isinstance(dr, AbstractDataRecord) or not dr.fixed_total_length:
-                raise ValueError(f"Incorrectly defined data structure for DID 0x{did:04X}. "
-                                 f"Only fixed length data records are supported right now.")
-            total_length += dr.min_occurrences * dr.length
-        return (RawDataRecord(name=name,
-                              children=data_records,
-                              length=total_length,
-                              min_occurrences=1,
-                              max_occurrences=1),)
-
-    return ConditionalFormulaDataRecord(formula=_get_did_data,
-                                        default_message_continuation=[default_did_data])
-
-
-def get_did_data_mask_2013(name: str, optional: bool) -> ConditionalFormulaDataRecord:
-    """
-    Get Conditional Data Record for DID data mask that is compatible with ISO 14229-1:2013 version.
-
-    :param name: Name for the Data Record that contains whole DID data mask.
-    :param optional: Whether the field is optional or mandatory.
-
-    :return: Conditional Data Record for DID data.
-    """
-    default_did_data_mask = RawDataRecord(name=name,
-                                          length=8,
-                                          min_occurrences=0 if optional else 1,
-                                          max_occurrences=None)
-
-    def _get_mask_data_record(data_record: AbstractDataRecord) -> RawDataRecord:
-        return MappingDataRecord(name=f"{data_record.name} (mask)",
-                                 length=data_record.length,
-                                 values_mapping={0: "no",
-                                                 data_record.max_raw_value: "yes"},
-                                 children=[_get_mask_data_record(child) for child in data_record.children],
-                                 min_occurrences=data_record.min_occurrences,
-                                 max_occurrences=data_record.max_occurrences)
-
-    def _get_did_data_mask(did: int) -> Tuple[RawDataRecord]:
-        data_records = DID_DATA_MAPPING_2013.get(did, None)
-        if data_records is None:
-            raise ValueError(f"No data structure defined for DID 0x{did:04X}.")
-        total_length = 0
-        mask_data_records = []
-        for dr in data_records:
-            if not isinstance(dr, AbstractDataRecord) or not dr.fixed_total_length:
-                raise ValueError(f"Incorrectly defined data structure for DID 0x{did:04X}. "
-                                 f"Only fixed length data records are supported right now.")
-            total_length += dr.min_occurrences * dr.length
-            mask_data_records.append(_get_mask_data_record(dr))
-        return (RawDataRecord(name=name,
-                              children=mask_data_records,
-                              length=total_length,
-                              min_occurrences=0 if optional else 1,
-                              max_occurrences=1),)
-
-    return ConditionalFormulaDataRecord(formula=_get_did_data_mask,
-                                        default_message_continuation=[default_did_data_mask])
-
-
-def get_did_data_mask_2020(name: str, optional: bool) -> ConditionalFormulaDataRecord:
-    """
-    Get Conditional Data Record for DID data mask that is compatible with ISO 14229-1:2020 version.
-
-    :param name: Name for the Data Record that contains whole DID data mask.
-    :param optional: Whether the field is optional or mandatory.
-
-    :return: Conditional Data Record for DID data.
-    """
-    default_did_data_mask = RawDataRecord(name=name,
-                                          length=8,
-                                          min_occurrences=0 if optional else 1,
-                                          max_occurrences=None)
-
-    def _get_mask_data_record(data_record: AbstractDataRecord) -> RawDataRecord:
-        return MappingDataRecord(name=f"{data_record.name} (mask)",
-                                 length=data_record.length,
-                                 values_mapping={0: "no",
-                                                 data_record.max_raw_value: "yes"},
-                                 children=[_get_mask_data_record(child) for child in data_record.children],
-                                 min_occurrences=data_record.min_occurrences,
-                                 max_occurrences=data_record.max_occurrences)
-
-    def _get_did_data_mask(did: int) -> Tuple[RawDataRecord]:
-        data_records = DID_DATA_MAPPING_2020.get(did, None)
-        if data_records is None:
-            raise ValueError(f"No data structure defined for DID 0x{did:04X}.")
-        total_length = 0
-        mask_data_records = []
-        for dr in data_records:
-            if not isinstance(dr, AbstractDataRecord) or not dr.fixed_total_length:
-                raise ValueError(f"Incorrectly defined data structure for DID 0x{did:04X}. "
-                                 f"Only fixed length data records are supported right now.")
-            total_length += dr.min_occurrences * dr.length
-            mask_data_records.append(_get_mask_data_record(dr))
-        return (RawDataRecord(name=name,
-                              children=mask_data_records,
-                              length=total_length,
-                              min_occurrences=0 if optional else 1,
-                              max_occurrences=1),)
-
-    return ConditionalFormulaDataRecord(formula=_get_did_data_mask,
-                                        default_message_continuation=[default_did_data_mask])
-
-
-def get_dids_2013(did_count: int,
-                  record_number: Optional[int],
-                  optional: bool = False) -> Tuple[Union[MappingDataRecord, ConditionalFormulaDataRecord], ...]:
-    """
-    Get DIDs related Data Records for given record (e.g. Snapshot or Stored Data).
-
-    .. note:: Supports only ISO 14229-1:2013 DIDs.
-
-    :param did_count: Number of DIDs that are part of the record that contains DIDs.
-    :param record_number: Order number of the record that contains DIDs.
-        None if this is the only DIDs group (e.g. part of ReadDataByIdentifier).
-    :param optional: False if the Data Record presence is mandatory, True otherwise.
-
-    :return: DIDs related Data Records that are part of the record.
-    """
-    data_records: List[Union[MappingDataRecord, ConditionalFormulaDataRecord]] = []
-    for did_number in range(did_count):
-        name = f"DID#{did_number + 1}" if record_number is None else f"DID#{record_number}_{did_number + 1}"
-        data_records.append(get_did_2013(name, optional=optional))
-        data_records.append(get_did_data_2013(name=f"{name} data"))
-    return tuple(data_records)
-
-
-def get_dids_2020(did_count: int,
-                  record_number: Optional[int],
-                  optional: bool = False) -> Tuple[Union[MappingDataRecord, ConditionalFormulaDataRecord], ...]:
-    """
-    Get DIDs related Data Records for given record (e.g. Snapshot or Stored Data).
-
-    .. note:: Supports only ISO 14229-1:2020 DIDs.
-
-    :param did_count: Number of DIDs that are part of the record that contains DIDs.
-    :param record_number: Order number of the record that contains DIDs.
-        None if this is the only DIDs group (e.g. part of ReadDataByIdentifier).
-    :param optional: False if the Data Record presence is mandatory, True otherwise.
-
-    :return: DIDs related Data Records that are part of the record.
-    """
-    data_records: List[Union[MappingDataRecord, ConditionalFormulaDataRecord]] = []
-    for did_number in range(did_count):
-        name = f"DID#{did_number + 1}" if record_number is None else f"DID#{record_number}_{did_number + 1}"
-        data_records.append(get_did_2020(name, optional=optional))
-        data_records.append(get_did_data_2020(name=f"{name} data"))
-    return tuple(data_records)
-
-
-def get_did_records_formula_2013(record_number: Optional[int]) -> Callable[[int], AliasMessageStructure]:
-    """
-    Get formula that can be used by Conditional Data Record for getting DID related Data Records.
-
-    .. note:: Supports only ISO 14229-1:2013 DIDs.
-
-    :param record_number: Order number of the record that contains DIDs.
-        None if this is the only DIDs group (e.g. part of ReadDataByIdentifier).
-
-    :return: Formula for given record (e.g. Snapshot or Stored Data).
-    """
-    return lambda did_count: get_dids_2013(did_count=did_count,
-                                           record_number=record_number)
-
-
-def get_did_records_formula_2020(record_number: Optional[int]) -> Callable[[int], AliasMessageStructure]:
-    """
-    Get formula that can be used by Conditional Data Record for getting DID related Data Records.
-
-    .. note:: Supports only ISO 14229-1:2020 DIDs.
-
-    :param record_number: Order number of the record that contains DIDs.
-        None if this is the only DIDs group (e.g. part of ReadDataByIdentifier).
-
-    :return: Formula for given record (e.g. Snapshot or Stored Data).
-    """
-    return lambda did_count: get_dids_2020(did_count=did_count,
-                                           record_number=record_number)
-
-
-def get_event_type_record_03_2013(event_number: Optional[int] = None) -> RawDataRecord:
-    """
-    Get eventTypeRecord Data Record for event equal to 0x03.
-
-    .. note:: Supports only ISO 14229-1:2013 DIDs.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventTypeRecord Data Record.
-    """
-    return RawDataRecord(name="eventTypeRecord" if event_number is None else f"eventTypeRecord#{event_number}",
-                         length=DID_BIT_LENGTH,
-                         children=(DID_2013,))
-
-
-def get_event_type_record_03_2020(event_number: Optional[int] = None) -> RawDataRecord:
-    """
-    Get eventTypeRecord Data Record for event equal to 0x03.
-
-    .. note:: Supports only ISO 14229-1:2020 DIDs.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventTypeRecord Data Record.
-    """
-    return RawDataRecord(name="eventTypeRecord" if event_number is None else f"eventTypeRecord#{event_number}",
-                         length=DID_BIT_LENGTH,
-                         children=(DID_2020,))
-
-
-def get_event_type_record_07_2013(event_number: Optional[int] = None) -> RawDataRecord:
-    """
-    Get eventTypeRecord Data Record for event equal to 0x07.
-
-    .. note:: Supports only ISO 14229-1:2013 DIDs.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventTypeRecord Data Record.
-    """
-    return RawDataRecord(name="eventTypeRecord" if event_number is None else f"eventTypeRecord#{event_number}",
-                         length=80,
-                         children=(DID_2013,
-                                   COMPARISON_LOGIC,
-                                   COMPARE_VALUE,
-                                   HYSTERESIS_VALUE,
-                                   LOCALIZATION))
-
-
-def get_event_type_record_07_2020(event_number: Optional[int] = None) -> RawDataRecord:
-    """
-    Get eventTypeRecord Data Record for event equal to 0x07.
-
-    .. note:: Supports only ISO 14229-1:2020 DIDs.
-
-    :param event_number: Order number of the event record to contain this Data Record.
-        None if there are no records.
-
-    :return: Created eventTypeRecord Data Record.
-    """
-    return RawDataRecord(name="eventTypeRecord" if event_number is None else f"eventTypeRecord#{event_number}",
-                         length=80,
-                         children=(DID_2020,
-                                   COMPARISON_LOGIC,
-                                   COMPARE_VALUE,
-                                   HYSTERESIS_VALUE,
-                                   LOCALIZATION))
+"""DID values mapping (compatible with ISO 14229-1:2013) to DID data message structure."""
 
 
 DID_2013 = MappingDataRecord(name="DID",
@@ -456,8 +104,3 @@ DATA_FROM_DID_2020 = RawDataRecord(name="Data from DID",
                                    min_occurrences=1,
                                    max_occurrences=None)
 
-EVENT_TYPE_RECORD_03_2013 = get_event_type_record_03_2013()
-EVENT_TYPE_RECORD_03_2020 = get_event_type_record_03_2020()
-
-EVENT_TYPE_RECORD_07_2013 = get_event_type_record_07_2013()
-EVENT_TYPE_RECORD_07_2020 = get_event_type_record_07_2020()
